requirements.txt
python-dotenv>=1.0.1
langchain
langgraph
langchain-openai
langchain-anthropic
langchain-google-genai
langchain-cohere
langchain-mistralai
langchain-huggingface>=0.0.7
numpy>=1.26.0
redis
pinecone>=6.0.0
requests>=2.32.4
tenacity>=9.0.0
pydantic>=2.7.0
pydantic-settings>=2.3.4
python-multipart>=0.0.18
aiofiles>=24.1.0
pytest
pytest-asyncio
httpx
pytest-mock
pytest-cov
codacy-coverage
coverage
black
isort
flake8
mypy
pre-commit
bandit
safety
boto3
openai
cohere
google-generativeai
mistralai
huggingface-hub
anthropic
transformers
accelerate
torch
sentence-transformers>=3.3.1
prometheus-client>=0.20.0
fastapi>=0.112.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.30
minio>=7.2.7
nats-py>=2.7.2
temporalio>=1.8.0
opentelemetry-api>=1.25.0
opentelemetry-sdk>=1.25.0
opentelemetry-instrumentation-fastapi>=0.46b0
opentelemetry-exporter-otlp>=1.25.0
psycopg[binary]>=3.2.1
pgvector>=0.2.5
# Pin to the latest available release (registry lists up to 1.16.5)
pygltflib==1.16.5
